src/main/java/com/springframework/recipe_app/service/IngredientServiceImpl.java
package com.springframework.recipe_app.service;

import com.springframework.recipe_app.commands.IngredientCommand;
import com.springframework.recipe_app.converters.IngredientToIngredientCommand;
import com.springframework.recipe_app.domain.Recipe;
import com.springframework.recipe_app.repositories.RecipeRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.util.Objects;
import java.util.Optional;

@Slf4j
@Service
public class IngredientServiceImpl implements IngredientService {

    private final RecipeRepository recipeRepository;
    private final IngredientToIngredientCommand ingredientToIngredientCommand;

    public IngredientServiceImpl(RecipeRepository recipeRepository, IngredientToIngredientCommand ingredientToIngredientCommand) {
        this.recipeRepository = recipeRepository;
        this.ingredientToIngredientCommand = ingredientToIngredientCommand;
    }

    @Override
    public IngredientCommand findByRecipeIdAndIngredientId(Long recipeId, Long ingredientId) {
        Optional<Recipe> recipeOptional = recipeRepository.findById(recipeId);

        if (recipeOptional.isEmpty()){
            log.error("recipe id not found. Id: {}", recipeId);
        }

        Recipe recipe = recipeOptional.get();

        Optional<IngredientCommand> ingredientCommand = recipe.getIngredients().stream()
                .filter(ingredient -> ingredient.getId().equals(ingredientId))
                .map(ingredientToIngredientCommand::convert)
                .filter(Objects::nonNull) // Ensure non-null values are kept
                .findFirst();

        if(ingredientCommand.isEmpty()){
            log.error("Ingredient id not found: {}", ingredientId);
        }

        return ingredientCommand.get();
    }
}
